# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NIPOPETO Pet Healthcare API",
    "version": "3.0.0",
    "description": "REST/JSON contract with bounded session/dashboard bootstrap and paginated collections. List GET response changed from array to {items,pagination,counts,related}. Default 20, maximum 100 rows."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
...
  "paths": {
...
    "/hotel-bookings": {
      "get": {
        "tags": [
          "Hotel"
        ],
        "summary": "List bookings",
        "operationId": "get_api_hotel_bookings",
        "responses": {
          "200": {
            "description": "Page of HotelBooking",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "pagination",
                    "counts",
                    "related"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/HotelBooking"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    "counts": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "integer",
                        "minimum": 0
                      }
                    },
                    "related": {
                      "$ref": "#/components/schemas/RelatedPageData"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Paginated JSON envelope. Rows use a stable order with id as a tie-breaker. Count, filters and page share a Repeatable Read transaction. related contains only relations needed by this page. Counts are independent of the selected status/page.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/SearchQuery"
          },
          {
            "$ref": "#/components/parameters/ExactId"
          },
          {
            "$ref": "#/components/parameters/OwnerIdFilter"
          },
          {
            "$ref": "#/components/parameters/PetIdFilter"
          },
          {
            "$ref": "#/components/parameters/StatusFilter"
          },
          {
            "name": "roomType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            },
            "description": "Optional database filter. Use all or omit for no filter."
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            },
            "description": "unbilled: eligible sources with no existing invoice."
          }
        ]
      },
      "post": {
        "tags": [
          "Hotel"
        ],
        "summary": "Create booking",
        "operationId": "post_api_hotel_bookings",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotelBookingEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotelBookingInput"
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 100,
              "pattern": "^[A-Za-z0-9_-]+$"
            },
            "description": "Giữ nguyên key khi gửi lại cùng yêu cầu; cùng key và payload khác trả 409."
          }
        ],
        "description": "Tạo một yêu cầu lưu trú chưa thanh toán. Chống gửi lặp và trùng ngày cho cùng thú cưng."
      }
    },
    "/hotel-bookings/{id}/status": {
      "patch": {
        "tags": [
          "Hotel"
        ],
        "summary": "Change booking status",
        "operationId": "patch_api_hotel_bookings_id_status",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotelBookingEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusInput"
              }
            }
          }
        },
        "description": "Checkout khách sạn chỉ khi hóa đơn cuối cùng đã paid; thao tác trạng thái cần expectedRevision."
      }
    },
    "/hotel-bookings/{id}/cancel": {
      "patch": {
        "tags": [
          "Hotel"
        ],
        "summary": "Cancel booking",
        "operationId": "patch_api_hotel_bookings_id_cancel",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotelBookingEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelInput"
              }
            }
          }
        }
      }
    },
    "/hotel-bookings/{id}/care-notes": {
      "post": {
        "tags": [
          "Hotel"
        ],
        "summary": "Add care note",
        "operationId": "post_api_hotel_bookings_id_care_notes",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CareNoteEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CareNoteInput"
              }
            }
          }
        }
      }
    },
...
    "/hotel-bookings/{id}/status-history": {
      "get": {
        "tags": [
          "Hotel bookings"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ],
        "operationId": "bookingStatusHistory",
        "summary": "Staff-only immutable status history; legacy history is not synthesized",
        "responses": {
          "200": {
            "description": "Events newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingStatusEvent"
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/RequestFailed"
          }
        }
      }
    },
    "/hotel-bookings/{id}/undo-status": {
      "post": {
        "tags": [
          "Hotel bookings"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ],
        "operationId": "undoHotelBookingStatus",
        "summary": "Reverse an eligible recorded check-in or reopen completion (admin only); reject dependent records and stale revisions",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "reason",
                  "expectedRevision"
                ],
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "expectedRevision": {
                    "type": "integer",
                    "minimum": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated booking, original event retained and reversal appended atomically",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "booking": {
                      "$ref": "#/components/schemas/HotelBooking"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/RequestFailed"
          }
        }
      }
    },
    "/hotel-bookings/{id}": {
      "get": {
        "tags": [
          "Hotel"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ],
        "operationId": "getHotelBookingDetail",
        "summary": "Xem chi tiết đặt phòng trong phạm vi chủ nuôi hoặc nhân viên",
        "responses": {
          "200": {
            "description": "Chi tiết đặt phòng",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotelBookingEnvelope"
                }
              }
            }
          },
          "default": {
            "description": "Yêu cầu thất bại",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
...
      "ApiError": {
        "description": "Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required or invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "RequestFailed": {
        "description": "Request failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 1000000,
          "default": 1
        },
        "description": "1-based page."
      },
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        },
        "description": "Rows per page; values above 100 return 422."
      },
      "SearchQuery": {
        "name": "q",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 100
        },
        "description": "Case-insensitive search applied in the database before pagination."
      },
      "ExactId": {
        "name": "id",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 128
        },
        "description": "Exact identifier lookup within the authenticated scope."
      },
      "OwnerIdFilter": {
        "name": "ownerId",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 128
        },
        "description": "Staff filter; owners are always scoped to their own JWT subject."
      },
      "ResourceId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "PetIdFilter": {
        "name": "petId",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 128
        },
        "description": "Optional database filter. Use all or omit for no filter."
      },
      "StatusFilter": {
        "name": "status",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 128
        },
        "description": "Optional database filter. Use all or omit for no filter."
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "doctor",
              "staff",
              "admin"
            ]
          },
          "address": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "fullName",
          "role"
        ]
      },
      "Pet": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "species": {
            "type": "string"
          },
          "breed": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "ageLabel": {
            "type": "string"
          },
          "weightKg": {
            "type": "number"
          },
          "microchipId": {
            "type": "string"
          },
          "healthStatus": {
            "type": "string"
          },
          "allergies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/String"
            }
          },
          "qrToken": {
            "type": "string"
          },
          "qrEnabled": {
            "type": "boolean"
          },
          "publicProfile": {
            "$ref": "#/components/schemas/PublicProfile"
          }
        },
        "required": [
          "id",
          "ownerId",
          "name",
          "species",
          "gender",
          "healthStatus",
          "allergies",
          "qrEnabled",
          "publicProfile"
        ]
      },
      "PublicProfile": {
        "type": "object",
        "properties": {
          "showOwnerPhone": {
            "type": "boolean"
          },
          "showOwnerEmail": {
            "type": "boolean"
          },
          "showOwnerAddress": {
            "type": "boolean"
          },
          "showMedicalAlerts": {
            "type": "boolean"
          },
          "rescueNote": {
            "type": "string"
          }
        },
        "required": [
          "showOwnerPhone",
          "showOwnerEmail",
          "showOwnerAddress",
          "showMedicalAlerts"
        ]
      },
      "Appointment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "petId": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "serviceName": {
            "type": "string"
          },
          "clinicName": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "time": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "doctorId": {
            "type": "string"
          },
          "ownerNote": {
            "type": "string"
          },
          "internalNote": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "statusRevision": {
            "type": "integer",
            "minimum": 0,
            "readOnly": true
          }
        },
        "required": [
          "id",
          "petId",
          "ownerId",
          "type",
          "serviceName",
          "date",
          "time",
          "status"
        ]
      },
      "MedicalRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "petId": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "appointmentId": {
            "type": "string"
          },
          "doctorName": {
            "type": "string"
          },
          "visitDate": {
            "type": "string",
            "format": "date"
          },
          "title": {
            "type": "string"
          },
          "diagnosis": {
            "type": "string"
          },
          "treatment": {
            "type": "string"
          },
          "weightKg": {
            "type": "number"
          },
          "temperatureC": {
            "type": "number"
          },
          "heartRateBpm": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "petId",
          "ownerId",
          "doctorName",
          "visitDate",
          "title",
          "diagnosis",
          "treatment"
        ]
      },
      "MedicalImage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "petId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "petId",
          "title",
          "imageUrl",
          "mimeType"
        ]
      },
      "HotelBooking": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "petId": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "checkIn": {
            "type": "string",
            "format": "date"
          },
          "checkOut": {
            "type": "string",
            "format": "date"
          },
          "nights": {
            "type": "integer"
          },
          "roomType": {
            "type": "string"
          },
          "serviceKeys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/String"
            }
          },
          "totalAmount": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "dailyCareNoteIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/String"
            }
          },
          "statusRevision": {
            "type": "integer",
            "minimum": 0,
            "readOnly": true
          }
        },
        "required": [
          "id",
          "petId",
          "ownerId",
          "checkIn",
          "checkOut",
          "nights",
          "roomType",
          "totalAmount",
          "status"
        ]
      },
      "CareNote": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "bookingId": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "note": {
            "type": "string"
          },
          "eatingStatus": {
            "type": "string"
          },
          "mood": {
            "type": "string"
          },
          "visibleToOwner": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "bookingId",
          "date",
          "note",
          "visibleToOwner"
        ]
      },
      "Notification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "recipientOwnerId": {
            "type": "string"
          },
          "recipientRole": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "actionUrl": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "recipientRole",
          "type",
          "title",
          "message",
          "status"
        ]
      },
...
      "String": {
        "type": "string"
      },
...
      "HotelBookingEnvelope": {
        "type": "object",
        "properties": {
          "booking": {
            "$ref": "#/components/schemas/HotelBooking"
          }
        },
        "required": [
          "booking"
        ]
      },
      "CareNoteEnvelope": {
        "type": "object",
        "properties": {
          "careNote": {
            "$ref": "#/components/schemas/CareNote"
          }
        },
        "required": [
          "careNote"
        ]
      },
...
      "StatusInput": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "internalNote": {
            "type": "string"
          },
          "expectedRevision": {
            "type": "integer",
            "minimum": 0,
            "description": "statusRevision from the booking displayed to the user; stale values return 409."
          }
        },
        "required": [
          "status",
          "expectedRevision"
        ]
      },
...
      "CancelInput": {
        "type": "object",
        "properties": {
          "ownerNote": {
            "type": "string"
          },
          "expectedRevision": {
            "type": "integer",
            "minimum": 0,
            "description": "statusRevision from the booking displayed to the user; stale values return 409."
          }
        },
        "required": [
          "expectedRevision"
        ]
      },
...
      "HotelBookingInput": {
        "type": "object",
        "properties": {
          "petId": {
            "type": "string"
          },
          "checkIn": {
            "type": "string",
            "format": "date"
          },
          "checkOut": {
            "type": "string",
            "format": "date"
          },
          "roomType": {
            "type": "string"
          },
          "serviceKeys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/String"
            }
          },
          "ownerNote": {
            "type": "string"
          }
        },
        "required": [
          "petId",
          "checkIn",
          "checkOut"
        ]
      },
      "CareNoteInput": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string"
          },
          "eatingStatus": {
            "type": "string"
          },
          "mood": {
            "type": "string"
          },
          "visibleToOwner": {
            "type": "boolean"
          }
        },
        "required": [
          "note"
        ]
      },
...
      "BookingStatusEvent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "bookingId": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "fromStatus": {
            "type": "string"
          },
          "toStatus": {
            "type": "string"
          },
          "actorId": {
            "type": "string"
          },
          "actorName": {
            "type": "string"
          },
          "actorRole": {
            "type": "string"
          },
          "revision": {
            "type": "integer"
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "reversesId": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Owner": {
        "allOf": [
          {
            "$ref": "#/components/schemas/User"
          },
          {
            "type": "object",
            "required": [
              "petIds",
              "loginEnabled"
            ],
            "properties": {
              "petIds": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "loginEnabled": {
                "type": "boolean",
                "description": "True only when this profile has login credentials."
              }
            }
          }
        ],
        "properties": {
          "petCount": {
            "type": "integer",
            "minimum": 0,
            "description": "Total pet count independent of pages. petIds is empty in bounded directory/profile responses; load pet pages with ownerId."
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ],
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "RelatedPageData": {
        "type": "object",
        "properties": {
          "owners": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Owner"
            }
          },
          "pets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Pet"
            }
          },
          "appointments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Appointment"
            }
          },
          "hotelBookings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HotelBooking"
            }
          },
          "medicalRecords": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MedicalRecord"
            }
          },
          "medicalImages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MedicalImage"
            }
          },
          "dailyCareNotes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CareNote"
            }
          },
          "notifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Notification"
            }
          },
          "invoices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Invoice"
            }
          }
        }
      }
    }
  }
}